build(backend; integration-tests): refactored integration test class names
