+++ diversity-lunch-be/src/test/java/de/brockhausag/diversitylunchspringboot/integrationstests/MatchingServiceIT.java
@@ -34,7 +34,7 @@ import static org.springframework.test.util.AssertionErrors.assertEquals;
 })
 @ActiveProfiles("Test")
 @Import(MsTeamsTestConfig.class)
-class MatchingServiceIntegrationTest {
+class MatchingServiceIT {
     @Autowired
     private MeetingRepository meetingRepository;
